Refactor argument reading in main.cpp
The argument reading has been refactored in main.cpp to provide more flexibility and functionality. This includes switching from standard input streams to an array of character pointers, allowing individual handling of arguments and enabling the use of a buffered stream. The changes also facilitate better tracking of short and long names and the count of sub-arguments.