Refactor CommandLineParser to use std::vector and Argument containers
The refactoring enhances code flexibility and allows argument parsing to use a vector in ArgumentContainer and Iterator instead of raw pointers. This change ensures a more modern and safer code design, as well as prepares the code base for potential future improvements such as parsing arguments with an arbitrary number of associated values.