Update command line parser and integrate std module
The code was updated to import the new `std` module and integrate it within the `cmd` module. This includes modifying the `CMakeLists.txt` to add `std` as an imported OBJECT library and adjust executable linking. Adjustments were done in `CommandLineParser.ixx` and `main.cpp` to use import instead of classic include for the `std` and `CommandLineParser`. Further, `main.cpp` was simplified by reducing the argument handling logic.