Add support for custom input stream reading
The code added to the program allows handling of a non-streamable class in which standard input streaming may not be applicable. By defining a custom_parameter_read function and utilizing the concept functionality of C++, the application can now process non-streamable classes without requiring the user to define explicit stream operator overloads.