Refactor command line parser to use iterators
In this commit, the command line parser is refactored to use iterator arguments instead of an argument container. This change simplifies the usage and makes the parser more flexible, allowing it to accept different types of inputs, as long as they fit the `string_like_ra_iterator` concept. Changes have been made in the parse, parse_sub_arguments, and parse_positionals methods, replacing the argument container with iterator inputs.